+++ R/app_server.R
@@ -2094,7 +2094,7 @@ app_server <- function(input, output, session) {
 
   # ______ SUMMARY RESULTS, results_overall _________ ####
   #. ####
-  # REPORT Overall - for BROWSER ####
+  # REPORT Overall/Multisite Summary - for BROWSER ####
 
   # *** The code here builds the html report shown in the shiny app but also see separate code used to download it.
 
@@ -2455,7 +2455,6 @@ app_server <- function(input, output, session) {
   })
   ####################################################### #
 
-
   #############################################################################  #
   ## * FOOTER ####
   ############################################ #
@@ -2467,7 +2466,6 @@ app_server <- function(input, output, session) {
   })
   #############################################################################  #
 
-
   if (isTRUE(getOption("shiny.testmode"))) {
     htmlwidgets::setWidgetIdSeed(12345) # ensures consistent element IDs across runs
     set.seed(12345)
@@ -2477,15 +2475,23 @@ app_server <- function(input, output, session) {
   #. ####
   #############################################################################  #
 
-  ##############################################  #
-  # REPORT Overall - for DOWNLOAD ####
+  downloadable_file_report_multisite <- reactive({
 
-  ### ejam2report() in Overall downloadHandler() ####
+    req(data_processed())
 
-  output$download_report_multisite <- downloadHandler(
+    # As soon as new results become available as updated data_processed(),
+    # this reactive re-runs, using snapshot of data run at that time
+    # so rendering is triggered by new results immediately, not by a download button,
+    # so HTML file is ready sooner for user when they click download button.
+    # And isolate() is used to avoid re-rendering if uploaded data or upload method menu changes,
+    # so changing anything but the title after running the results
+    # will not re-render this report
 
-    filename = function() {
-      create_filename(
+    analysis_title <- sanitized_analysis_title()
+
+    isolate({
+
+      filename <- create_filename(
         file_desc = 'community report',
         title =  sanitized_analysis_title(),
         buffer_dist = submitted_radius_val(),
@@ -2493,24 +2499,21 @@ app_server <- function(input, output, session) {
         with_datetime = TRUE,
         ext = ifelse(input$format1pager == 'pdf', '.pdf', '.html')
       )
-    },
-    content = function(file) {
-
-      # report_fname(file) # never used now?
+      filename_fullpath <- file.path(tempdir(), filename)
 
       if (submitted_upload_method() == 'SHP') {
         shp_for_report <- data_uploaded()
       } else {
         shp_for_report <- NULL
       }
-      ejam2report(
+      report_path <- ejam2report(
         fileextension = '.html',
-        filename = file,
+        filename = filename_fullpath,
         ejamitout = data_processed(),
         return_html = FALSE,
         submitted_upload_method = submitted_upload_method(),
         shp = shp_for_report,
-        analysis_title =  sanitized_analysis_title(),
+        analysis_title =  analysis_title,
         launch_browser = FALSE,
         show_ratios_in_report = isTRUE(as.logical(input$show_ratios_in_report)),
         extratable_show_ratios_in_report = isTRUE(as.logical(input$extratable_show_ratios_in_report)),
@@ -2521,8 +2524,25 @@ app_server <- function(input, output, session) {
         logo_path =  EJAM:::global_or_param("report_logo"), # use FULL path for ejam2report() unlike for UI build # app_sys("report/community_report/ejamhex4.png"),   # NULL means default, "" means no logo
         logo_html = NULL # this is the report logo, NOT app_logo_html... and gets defined downstream based on logo_path
       )
+    })
+    if (input$testing) {
+    cat(paste0('in reactive creating report file, filename created as param is: \n"', filename, '"\n'))
+    cat(paste0('in reactive creating report file, report_path created by ejam2report() is: \n"', report_path, '"\n'))
+    }
+    report_path
+  })
+  ##############################################  #
+  # REPORT Overall/Multisite Summary - for DOWNLOAD ####
 
+  ### ejam2report() in Overall downloadHandler() ####
 
+  output$download_report_multisite <- downloadHandler(
+
+    filename = function() {
+      basename(downloadable_file_report_multisite())
+    },
+    content = function(file) {
+      file.copy(from = downloadable_file_report_multisite(), to = file, overwrite = T)
     }
   )
   #############################################################################  #
@@ -2621,7 +2641,7 @@ cat("Clicked on site #", sitenumber, "for a 1-site report\n")
           title = "Download Ready",
           "Your report is ready. Click the button below to download.",
           footer = tagList(
-            downloadButton("download_report_single_site", "Download Report"),
+            downloadButton('download_report_single_site', label = 'Download Report'),
             modalButton("Close")
           ),
           easyClose = TRUE,
@@ -2713,35 +2733,8 @@ cat("Clicked on site #", sitenumber, "for a 1-site report\n")
 
       html_content <- isolate({
 
-        temp_file <- tempfile(fileext = '.html')
+        downloadable_file_report_multisite() # already was created and saved in tempfile here
 
-        if (submitted_upload_method() == 'SHP') {
-          shp_for_report <- data_uploaded()
-        } else {
-          shp_for_report <- NULL
-        }
-
-        # *** ejam2report() or this server code should be made to handle FIPS upload and FIPS dropdown
-        # & use obtained FIPS polygons + analysis results as shp
-        ## note the mapping code in server already does that, creating report_map()
-
-        ejam2report(
-
-          fileextension = '.html',
-          filename = temp_file,
-          ejamitout = data_processed(),
-          return_html = FALSE,
-          submitted_upload_method = submitted_upload_method(),
-          shp = shp_for_report,
-          analysis_title =  sanitized_analysis_title(),
-          launch_browser = FALSE,
-          show_ratios_in_report = isTRUE(as.logical(input$show_ratios_in_report)),
-          extratable_show_ratios_in_report = isTRUE(as.logical(input$extratable_show_ratios_in_report)),
-          extratable_title = input$extratable_title,
-          extratable_title_top_row = input$extratable_title_top_row,
-          extratable_list_of_sections = EJAM:::global_or_param("default_extratable_list_of_sections"),
-          extratable_hide_missing_rows_for = input$extratable_hide_missing_rows_for
-        )
       })
 
       ## ejam2excel() ####